packages/backend/requirements.txt
flask==3.0.3
gunicorn==22.0.0
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
flask-jwt-extended==4.6.0
redis==5.0.6
pydantic==2.7.4
pydantic-settings==2.4.0
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
apscheduler==3.10.4
requests==2.32.3
pypdf==4.3.1
openai==1.37.1
twilio==9.3.2
pytest==8.2.2
black==24.8.0
flake8==7.0.0
bandit==1.7.9

